Various texture mapping related changes:
- Add sampler2D support to native C++ shader builder
- Auto-load PNGs on first use in rasterizer material activation
- Auto-load PNGs on first use in native C++ shader builder
- Add texture2d support to native C++ shader builder
- Add lightgradient support to native C++ shader builder
- Add lightgradient support to GLSL shader builder
- Remove hard-coded LightGradient material from ray-tracer
- Move ray-tracer parsers into their own file
- Misc ray-tracer code clean-up
- Move geometry collection into SpatialIndex class in ray-tracer
- Add screenshot functionality to ray-tracer

Other notes: the auto-load functionality in both the shader builders is of questionable design.  Right now, it occurs in the rasterizer for the GLSL version, but in the builder for the native C++ version, which is an instant sign the design is not right.  This is an incremental step which hopefully will evolve into a more coherent design as the notion of a full texture cache is developed for LxEngine.